requirements.txt
yfinance>=0.2.0
pyyaml>=6.0
python-dotenv>=1.0.0
pandas>=2.0.0
zai-sdk>=0.1.0
